+++ pkg/injector/injector_test.go
@@ -54,35 +54,6 @@ func TestConfigCorrectValues(t *testing.T) {
 	assert.Equal(t, "e", injector.config.Namespace)
 }
 
-func TestGetServiceAddress(t *testing.T) {
-	testCases := []struct {
-		name          string
-		namespace     string
-		clusterDomain string
-		port          int
-		expect        string
-	}{
-		{
-			port:          80,
-			name:          "a",
-			namespace:     "b",
-			clusterDomain: "cluster.local",
-			expect:        "a.b.svc.cluster.local:80",
-		},
-		{
-			port:          50001,
-			name:          "app",
-			namespace:     "default",
-			clusterDomain: "selfdefine.domain",
-			expect:        "app.default.svc.selfdefine.domain:50001",
-		},
-	}
-	for _, tc := range testCases {
-		dns := getServiceAddress(tc.name, tc.namespace, tc.clusterDomain, tc.port)
-		assert.Equal(t, tc.expect, dns)
-	}
-}
-
 func TestAnnotations(t *testing.T) {
 	t.Run("Config", func(t *testing.T) {
 		m := map[string]string{annotations.KeyConfig: "config1"}
