--- pkg/injector/pod_patch_test.go
@@ -1,125 +1,0 @@
-/*
-Copyright 2022 The Dapr Authors
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package injector
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	corev1 "k8s.io/api/core/v1"
-
-	"github.com/dapr/dapr/pkg/injector/annotations"
-)
-
-func TestPodContainsVolume(t *testing.T) {
-	testCases := []struct {
-		testName   string
-		podVolumes []corev1.Volume
-		volumeName string
-		expect     bool
-	}{
-		{
-			"pod with no volumes",
-			[]corev1.Volume{},
-			"volume1",
-			false,
-		},
-		{
-			"pod does not contain volume",
-			[]corev1.Volume{
-				{Name: "volume"},
-			},
-			"volume1",
-			false,
-		},
-		{
-			"pod contains volume",
-			[]corev1.Volume{
-				{Name: "volume1"},
-				{Name: "volume2"},
-			},
-			"volume2",
-			true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.testName, func(t *testing.T) {
-			pod := corev1.Pod{
-				Spec: corev1.PodSpec{
-					Volumes: tc.podVolumes,
-				},
-			}
-			assert.Equal(t, tc.expect, podContainsVolume(pod, tc.volumeName))
-		})
-	}
-}
-
-func TestGetVolumeMounts(t *testing.T) {
-	testCases := []struct {
-		testName                  string
-		volumeReadOnlyAnnotation  string
-		volumeReadWriteAnnotation string
-		podVolumeMountNames       []string
-		expVolumeMounts           []corev1.VolumeMount
-	}{
-		{
-			"no annotations",
-			"",
-			"",
-			[]string{"mount1", "mount2"},
-			[]corev1.VolumeMount{},
-		},
-		{
-			"annotations with volumes present in pod",
-			"mount1:/tmp/mount1,mount2:/tmp/mount2",
-			"mount3:/tmp/mount3,mount4:/tmp/mount4",
-			[]string{"mount1", "mount2", "mount3", "mount4"},
-			[]corev1.VolumeMount{
-				{Name: "mount1", MountPath: "/tmp/mount1", ReadOnly: true},
-				{Name: "mount2", MountPath: "/tmp/mount2", ReadOnly: true},
-				{Name: "mount3", MountPath: "/tmp/mount3", ReadOnly: false},
-				{Name: "mount4", MountPath: "/tmp/mount4", ReadOnly: false},
-			},
-		},
-		{
-			"annotations with volumes not present in pod",
-			"mount1:/tmp/mount1,mount2:/tmp/mount2",
-			"mount3:/tmp/mount3,mount4:/tmp/mount4",
-			[]string{"mount1", "mount2", "mount4"},
-			[]corev1.VolumeMount{
-				{Name: "mount1", MountPath: "/tmp/mount1", ReadOnly: true},
-				{Name: "mount2", MountPath: "/tmp/mount2", ReadOnly: true},
-				{Name: "mount4", MountPath: "/tmp/mount4", ReadOnly: false},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.testName, func(t *testing.T) {
-			pod := corev1.Pod{}
-			pod.Annotations = map[string]string{
-				annotations.KeyVolumeMountsReadOnly:  tc.volumeReadOnlyAnnotation,
-				annotations.KeyVolumeMountsReadWrite: tc.volumeReadWriteAnnotation,
-			}
-			pod.Spec.Volumes = []corev1.Volume{}
-			for _, volumeName := range tc.podVolumeMountNames {
-				pod.Spec.Volumes = append(pod.Spec.Volumes, corev1.Volume{Name: volumeName})
-			}
-
-			volumeMounts := getVolumeMounts(pod)
-			assert.Equal(t, tc.expVolumeMounts, volumeMounts)
-		})
-	}
-}
